rename namespace in DurabilityExtensions.cs
