+++ src/SanchozzONIMods.sln
@@ -19,6 +19,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	ProjectSection(SolutionItems) = preProject
 		lib\BaseConfig.cs = lib\BaseConfig.cs
 		lib\BaseOptions.cs = lib\BaseOptions.cs
+		lib\DurabilityExtensions.cs = lib\DurabilityExtensions.cs
 		lib\EffectsExtensions.cs = lib\EffectsExtensions.cs
 		lib\FXLoopAnim.cs = lib\FXLoopAnim.cs
 		lib\StorageDropper.cs = lib\StorageDropper.cs
