+++ src/SuitRecharger/SuitRecharger.csproj
@@ -20,14 +20,14 @@
     <UsesPLib>true</UsesPLib>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="..\lib\DurabilityExtensions.cs">
+      <Link>DurabilityExtensions.cs</Link>
+    </Compile>
     <Compile Include="..\lib\StorageDropper.cs">
       <Link>StorageDropper.cs</Link>
     </Compile>
     <Compile Include="..\lib\Utils.cs">
       <Link>Utils.cs</Link>
-    </Compile>
-    <Compile Include="..\WornSuitDischarge\DurabilityExtensions.cs">
-      <Link>DurabilityExtensions.cs</Link>
     </Compile>
     <Compile Include="AlwaysFunctionalConduitDispenser.cs" />
     <Compile Include="SuitRecharger.cs" />
