+++ src/WornSuitDischarge/WornSuitDischarge.csproj
@@ -20,7 +20,9 @@
     <UsesPLib>true</UsesPLib>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="DurabilityExtensions.cs" />
+    <Compile Include="..\lib\DurabilityExtensions.cs">
+      <Link>DurabilityExtensions.cs</Link>
+    </Compile>
     <Compile Include="WornSuitDischargePatches.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
